% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_to_decimal.R
\name{bond_to_decimal}
\alias{bond_to_decimal}
\title{Converts a vector of fractional bond prices to decimal.}
\usage{
bond_to_decimal(bond_px)
}
\arguments{
\item{bond_px}{A character vector of fractional bond prices, can include decimal prices too}
}
\value{
Returns a vector of decimalized bond prices
}
\description{
This function is designed to handle all flavors of
fixed income fractional price quotes and to return decimal equivalents
for use in your scripts. Requires a character vector input.
}
\examples{

# convert fractional bond prices to decimal prices
x <- c("$101-28", "101-2 1/8", "99+")
bond_to_decimal(x)

}
