% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_cusip.R
\name{validate_cusip}
\alias{validate_cusip}
\title{Checks for CUSIP validity.}
\usage{
validate_cusip(cusips)
}
\arguments{
\item{cusips}{A character vector of presumed CUSIP identifiers}
}
\value{
Returns a logical vector of CUSIP validity
}
\description{
Using the CUSIP Global services methodology, this function checks
the validity of CUSIPs. The check relies on the Modulus 10 Double Add
Double algorithm. It does not query to check the assignment of any identifier
only its compositional validity. This is useful for checking hand or OCR
generated input. All non-9 digit inputs will return as FALSE (i.e. invalid
cusip).
}
\examples{

# check vector of CUSIPs for validity
x <- c("037833100", "0378331O0")
validate_cusip(x)

}
