% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_to_numeric.R
\name{rating_to_numeric}
\alias{rating_to_numeric}
\title{Converts a vector of bond credit ratings to numeric values}
\usage{
rating_to_numeric(
  credit_rating,
  nr_as_na = TRUE,
  summary_fun = NULL,
  ...,
  summarize_numeric = FALSE
)
}
\arguments{
\item{credit_rating}{A character vector of bond credit ratings}

\item{summary_fun}{Function input to summarize vector}

\item{...}{Additional arguments passed to summary_fun}

\item{summarize_numeric}{Return a numeric summary value.
Defaults to FALSE to return letter rating equivalent.}

\item{na_as_nr}{Assign missing or unrecognized ratings highest numeric value,
defaults to TRUE}
}
\value{
Returns a vector of numeric credit rating equivalents.
}
\description{
This function takes a vector of bond credit ratings and
returns a vector of their numeric equivalents or, optionally, a statistical
summary of the translated values. The package currently supports
long term credit ratings from Moody's, S&P, Fitch, Kroll, &
DBRS/Morningstar. All mappings are stored in creditor:::cr_imp. Short term
ratings are assigned the floor rounded value of their long term equivalents.
By default, NA and unrecognized ratings are not assigned a value. This setting
can be changed to assign the highest numeric value (23) to missing or unrated
bonds.
}
\examples{

# convert bond credit ratings to numeric values
x <- c("AAA", "Baa2", "A(low)")
rating_to_numeric(x)

# summarize ratings
x <- c("AAA", "Baa2", "A(low)")
rating_to_numeric(x, summary_fun = weighted.mean, w = c(0.15, 0.60, 0.25))
}
